Don't remove subscription if the old is the same as the new

+++ wp-web-push/wp-web-push-main.php
@@ -126,7 +126,7 @@ class WebPush_Main {
   }
 
   public static function handle_register() {
-    if (isset($_POST['oldEndpoint'])) {
+    if (isset($_POST['oldEndpoint']) && $_POST['oldEndpoint'] !== $_POST['endpoint']) {
       WebPush_DB::remove_subscription($_POST['oldEndpoint']);
     }
